django_photo_management/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2
django-simple-captcha>=0.6.0
openpyxl>=3.1.0
celery==5.3.4
redis==5.0.1
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
django-celery-results==2.5.1
django-celery-beat==2.5.0
oracledb>=2.0.0
cryptography>=42.0.0